Combine s3 and download submodules into one
